# Here's your landing page
**I recommend clicking that big, exciting "fork" button in the upper-right.** Then you'll own the code.

You'll notice we're on a special `gh-pages` branch, meaning that the contents of this repository are available at `http://appstop.github.io/<this-repo-name>`. *It may take a while for Github to make the page available.* After you fork it, it'll be available at `http://<your-username>.github.io/<this-repo-name>`.

Once you've forked this repository into your own account, you can modify the HTML or set up and a custom domain using the instructions [here](https://help.github.com/articles/setting-up-a-custom-domain-with-github-pages/).